Handle decorated pots
Since 1.20.3, decorated pots have 1 stack of inventory. Handle events
for depositing them and breaking them with an arrow. Adds decorated pots
to the default config.yml
